convert mutations to actions

+++ packages/core/epics/dbEpics.ts
@@ -1,7 +1,8 @@
 import { Account, Bank, Bill, Budget, Category, Db, Setting, Transaction } from '@ag/db/entities'
 import crypto from 'crypto'
+import { defineMessages } from 'react-intl'
 import { from, of } from 'rxjs'
-import { catchError, filter, mergeMap, switchMap, withLatestFrom } from 'rxjs/operators'
+import { catchError, filter, map, mergeMap, switchMap, withLatestFrom } from 'rxjs/operators'
 import sanitize from 'sanitize-filename'
 import { isActionOf } from 'typesafe-actions'
 import { actions } from '../actions'
@@ -63,7 +64,7 @@ const dbCreateEpic: CoreEpic = (action$, state$, { sys: { openDb } }) =>
         const connection = await openDb(dbInfo.path, key, appEntities)
         await dbRepo.save(dbInfo)
 
-        const dbs: DbInfo[] = await dbRepo.find()
+        const dbs = await dbRepo.all()
 
         return { connection, dbs }
       })()
@@ -88,7 +89,7 @@ const dbOpenEpic: CoreEpic = (action$, state$, { sys: { openDb } }) =>
       const dbRepo = selectors.getDbRepository(state)
 
       const p = (async () => {
-        const dbInfo = await dbRepo.findOneOrFail(dbId)
+        const dbInfo = await dbRepo.get(dbId)
         const key = dbInfo.getKey(password)
 
         const connection = await openDb(dbInfo.path, key, appEntities)
@@ -105,35 +106,48 @@ const dbOpenEpic: CoreEpic = (action$, state$, { sys: { openDb } }) =>
     })
   )
 
-const dbDeleteEpic: CoreEpic = (action$, state$, { sys: { deleteDb } }) =>
+const dbDeleteEpic: CoreEpic = (action$, state$, { ui: { alert, showToast }, sys: { deleteDb } }) =>
   action$.pipe(
-    filter(isActionOf(actions.dbDelete.request)),
+    filter(isActionOf(actions.deleteDb)),
     withLatestFrom(state$),
     switchMap(([action, state]) => {
       const { dbId } = action.payload
       const dbRepo = selectors.getDbRepository(state)
+      const intl = selectors.getIntl(state)
 
-      const p = (async () => {
-        const dbInfo = await dbRepo.findOneOrFail(dbId)
-        await deleteDb(dbInfo.path)
+      const prompt = () =>
+        new Promise<boolean>((resolve, reject) => {
+          alert({
+            title: intl.formatMessage(messages.title),
+            body: intl.formatMessage(messages.body),
+            danger: true,
 
-        await dbRepo
-          .createQueryBuilder()
-          .delete()
-          .from(Db)
-          .where('dbId = :dbId', { dbId })
-          .execute()
+            confirmText: intl.formatMessage(messages.delete),
+            onConfirm: () => resolve(true),
 
-        const dbs: DbInfo[] = await dbRepo.find()
+            cancelText: intl.formatMessage(messages.cancel),
+            onCancel: () => resolve(false),
+          })
+        })
 
-        return { dbs }
+      const p = (async (): Promise<DbInfo[] | undefined> => {
+        const confirmed = await prompt()
+
+        if (confirmed) {
+          const dbInfo = await dbRepo.get(dbId)
+          await deleteDb(dbInfo.path)
+          await dbRepo.deleteDb(dbId)
+
+          showToast(intl.formatMessage(messages.deleted), true)
+
+          const dbs = await dbRepo.all()
+          return dbs
+        }
       })()
 
       return from(p).pipe(
-        mergeMap(res => [
-          actions.dbDelete.success(res), //
-        ]),
-        catchError(error => of(actions.dbDelete.failure(error)))
+        filter((res): res is DbInfo[] => !!res),
+        map(res => actions.dbSetInfos(res)) //
       )
     })
   )
@@ -144,3 +158,26 @@ export const dbEpics = [
   dbOpenEpic,
   dbDeleteEpic,
 ]
+
+const messages = defineMessages({
+  title: {
+    id: 'dbEpics.title',
+    defaultMessage: 'Are you sure?',
+  },
+  body: {
+    id: 'dbEpics.body',
+    defaultMessage: 'This will all your data.  This action cannot be undone.',
+  },
+  delete: {
+    id: 'dbEpics.delete',
+    defaultMessage: 'Delete',
+  },
+  cancel: {
+    id: 'dbEpics.cancel',
+    defaultMessage: 'Cancel',
+  },
+  deleted: {
+    id: 'dbEpics.deleted',
+    defaultMessage: 'Data deleted',
+  },
+})
